refactor(auth): change the background to the same widget
