+++ lib/src/page/pages.dart
@@ -1,4 +1,4 @@
 export 'init_page.dart';
 export 'main_page.dart';
-export 'login_page.dart';
-export 'signup_page.dart';
+export 'auth/login_page.dart';
+export 'auth/signup_page.dart';
